Remove not strictly necessary fields from `game_data::Item` & fix collectable handling
The fix is implemented by baking the CL character into the names of items that are always collectable. These changes fix all major issues of the last commit but actually slightly increases the binary size due to the baked CL characters and maybe some alignment issues (size of the wasm increases by roughly ~10kB). Additionally, any code written in the future that works with recipe item names must account for the baked character.